Fix issue when commenting on new files.
When a pull request adds a new file and it also has other
commits that modified that new added file, linterbot fails to
comment because it was using the line number suggested by swiftlint
. That is only valid if the new file is added in only one commit
and there are not any other commits modifying that file. If that
is not the case then the violation must be matched in the commits'
patch that added the violation.